Open and Open with actions
Implementation of Open and Open with actions. This create a custom
dialog that allow show all registered application even when only one
application exists or has a preferred application.
This change requires a platform signature to allow use
addPreferredActivity or removePreferredActivity.
For this is not working. Need be tested and checked inside CM.