Don't the method onSaveInstanceState and onRestoreInstanceState
NavigationActivity is singleTop, so when it is displayed no
other activities exists in the stack. This cause an issue with the saved
instance of this class, because if another activity is displayed, and
the process is killed, NavigationActivity is started and the saved
instance gets corrupted. For this reason the methods onSaveInstanceState
and onRestoreInstanceState are not implemented, and every time the app
is killed, is restarted from his initial state.