Jail room mode
This is the mode by default of the application. This change create an
non-privileged space (like a jail) from which the user can not exit,
This zone is the storage volumes.
This hides all advanced functionalities that could give access to
another filesystem zones and actions that could damage the system.

* CleanUp